refactor(AuthorizationException): 권한 예외를 ui 패키지로 이동
