+++ src/main/java/wooteco/subway/auth/ui/AuthorizationException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.auth.application;
+package wooteco.subway.auth.ui;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
